requirements.txt
SpeechRecognition>=3.10.0
PyAudio>=0.2.14
pyttsx3>=2.90
pyautogui>=0.9.54
selenium>=4.15.0
webdriver-manager>=4.0.1
beautifulsoup4>=4.12.0
requests>=2.31.0
lxml>=5.0.0
anthropic>=0.80.0
google-genai>=1.0.0
groq>=0.9.0
python-dotenv>=1.0.0
